only download new mhtmls from s3

+++ bananalyzer/__main__.py
@@ -319,11 +319,12 @@ def main() -> int:
 
     for example in examples:
         if example.mhtml_url is not None:
-            mhtml_str = download_mhtml(example.mhtml_url)
             mhtml_path = get_examples_path() / example.id / "index.mhtml"
-            mhtml_path.parent.mkdir(parents=True, exist_ok=True)
-            with open(mhtml_path, "w") as file:
-                file.write(mhtml_str)
+            if not mhtml_path.exists():
+                mhtml_path.parent.mkdir(parents=True, exist_ok=False)
+                mhtml_str = download_mhtml(example.mhtml_url)
+                with open(mhtml_path, "w") as file:
+                    file.write(mhtml_str)
 
     # Load the desired tests
     generator = PytestTestGenerator()
